services/patient_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
